requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
emoji==2.14.0
fastapi==0.115.2
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.52.0
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
RapidFuzz==3.10.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.11.0
starlette==0.40.0
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
webrtcvad==2.0.10
websocket-client==1.8.0
